Rename to EmbeddedChart
