+++ apps/charterafrica/src/components/EmbeddedChart/index.js
@@ -1,0 +1,3 @@
+import EmbeddedChart from "./EmbeddedChart";
+
+export default EmbeddedChart;
