--- apps/charterafrica/src/components/FlourishCharts/index.js
@@ -1,3 +1,0 @@
-import FlourishCharts from "./FlourishCharts";
-
-export default FlourishCharts;
